--- src/components/EmployeeDetails/EmployeeSheetDisplay/EmployeeSheetDisplay.module.css
@@ -1,77 +1,0 @@
-.profile {
-    margin: 5rem auto;
-    text-align: center;
-  }
-  
-  .profile h1 {
-    font-size: 5rem;
-  }
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- 
- /* .tablecontainer{
-    b
-    
-    
-    order-collapse: collapse;
-    margin: 25px 0;
-    font-size: 0.9em;
-    min-width: 400px; */
-    /* background-image: url("../../../assets/images/foremanbg.jpg"); */
-    /* padding: 2% 3%;
-    width: 100%;
-    height: 100vh;
-    align-items: center; */
-    /* box-sizing: border-box; */
-    /* background-color: rgb(209, 211, 211);
-    box-shadow: 0 0px 10px rgba(0, 0, 0, 0.75);
-    margin-right: 30%; */
-    /* margin: 25px 0; */
-    /* font-size: 0.9em; */
-    /* min-width: 900px;
-    border-collapse: collapse; */
-    /* margin: auto; */
- /*  */
- /* } */
- 
- /* .tablecontainer .table{
-    background-color: rgb(255, 255, 255);
-    border-radius: 30px;
-    margin-top: 1%;
-    width: 30%;
-    margin: auto;
- }
- 
- .tablehead{
-    background-color: rgb(188, 188, 188);
-    position: sticky;
-    top: 0px; */
-    /* padding: 2% 3%; */
- /* }
- 
- .tablebody{
-    background-color: rgb(255, 255, 255);
- } */
-
