--- src/pages/BillGeneration/invoice.module.css
@@ -1,24 +1,0 @@
-.heading{
-    color: black;
-    margin-left: 500px;
-}
-.jobcard {
-    margin: 5rem auto;
-    text-align: center;
-  }
-  .btn{
-    cursor: pointer;
-  padding: 70px;
-  }
-  
-
-.stylingtable, .stylingheader, .stylinrow{
-    border: 1px solid;
-    width:30%;
-    margin-left: 30px;
-    
-    
-
-}
- 
-
