--- src/pages/StartingPage/StartingPageContent.module.css
@@ -1,93 +1,0 @@
-.starting {
-  /* margin: 3rem auto; */
-  text-align: center;
-  background-image: url("https://img.freepik.com/premium-photo/future-building-construction-engineering-project_31965-2731.jpg?size=626&ext=jpg&ga=GA1.2.1228466384.1667154480&semt=sph");   
-  /* ../images/landingpage.jpg */
-  width: 100%;
-  height: 100vh;
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: cover;
-  /* margin-top: 0px; */
-  margin-bottom: 0px;
-  
-}
-
-.starting h1 {
-  font-size: 5rem;
-}
-
-
-  /* background-color:rgb(214, 214, 209); */
-  /* height: 200px; */
-  /* width: 400px;
-  text-align: center; */
-  /* margin-left: 400px;
-  padding-top: 40px;
-  margin-top: 20px; */
-  /* border: none;
-  border-radius: 45px; */
-  
-
-
-.about h2 {
-  font-size: 45px;
-  color: rgb(0, 0, 0);
-  padding-right: 50%;
-  padding-top: 15%;
-  margin: 0 0;
-}
-.about p {
-  font-size: 20px;
-  color: rgb(0, 0, 0);
-  padding-top: 1%;
-  width: 500px;
-  font-weight: bold;
-  /* padding-top: 20%; */
-  margin-right: 50%;
-  padding-left: 150px;
-
- 
-}
-
-.about button {
-  /* padding-right: 50%; */
-
-
-}
-
-.heading {
-  text-align: center;
-  padding:2.5rem 0;
-  /* background-color: antiquewhite; */
-  
-
-}
-
-.heading span{
-  font-size: 3.5rem;
-  /* background:rgba(255, 165, 0,.2); */
-  /* color:var(--orange); */
-  color: var(--green);
-  border-radius: .5rem;
-  padding:.2rem 1rem;
-}
-
-.btn{
-  display: inline-block;
-  margin-top: 1rem;
-  background: rgba(0, 0, 0, 0.705);
-  color:#fff;
-  padding:.8rem 3rem;
-  border:.2rem solid black;
-  cursor: pointer;
-  font-size: 1.0rem;
-  border-radius: 8px;
-  margin-right: 50%;
-}
-
-.btn:hover{
-  /* background:silver; */
-  color:rgb(3, 140, 128);
-  border: 0.1px solid;
-}
